hidgd: add TPM functions to give correct registration
Use the TPM to obtain a static EC key at 81000101 as the public point
and signature for registration.  We assume the DER certificate is a
signed version of this.  Currently, I'm just using a self signed
certificate which the webauthn demo site seems to accept.

https://webauthn.org/

This commit of code passes registration, but obviously not login
because signing is currently unimplemented.

Signed-off-by: James Bottomley <James.Bottomley@HansenPartnership.com>