hidgd: add correct AUTHENTICATE response
The key handle now contains the TPM representation of an elliptic
curve key, so unpack this key and use it to sign the incoming
challenge.  This scheme is now sufficient to pass the
https://webauthn.org test for both registration and login.  However,
the counter is ephemeral to the hidgd so we need a permanent solution
for that as well.

Signed-off-by: James Bottomley <James.Bottomley@HansenPartnership.com>